Extract registry tests to a test case module
To allow reuse by external registry providers who want to implement the `Commanded.Registration` behaviour.
