+++ test/registration/local_registry_test.exs
@@ -1,0 +1,6 @@
+defmodule Commanded.LocalRegistryTest do
+  alias Commanded.RegistrationTestCase
+  alias Commanded.Registration.LocalRegistry
+
+  use RegistrationTestCase, registry: LocalRegistry
+end
